passenger refactored

+++ src/main/java/com/dmytrobozhor/airlinereservationservice/service/AbstractPassengerService.java
@@ -4,7 +4,7 @@ import com.dmytrobozhor.airlinereservationservice.domain.Passenger;
 
 import java.util.Optional;
 
-public interface AbstractPassengerService extends AbstractCrudService<Passenger, Integer> {
+public interface AbstractPassengerService extends AbstractCrudService<Passenger, Long> {
 
     Optional<Passenger> findByPhoneNumber(String phoneNumber);
